Move parsePncDate function into root lib/parse folder
